MoveSelector splitted
